cloud: Move Cloud install logic into Model CLI and add GPU detection
After prototyping the middleware-based Cloud approach, it's become clear
that there's so much overlap with the Docker CE support that we should
just move the installation into the Model CLI. We now no longer need
middleware in the Cloud CLI.

This required adding GPU detection, which is a little more flexible now
and also applies to Docker CE.

This does put us in the position of having a Cloud engine's runner's TCP
port exposed locally, but that was the case already for Cloud due to TCC
implmentation details. We still have time to refine this transport.

Overall, this keeps the logic much simpler and doesn't require
duplication in the Cloud CLI codebase.

Signed-off-by: Jacob Howard <jacob.howard@docker.com>