Return error from the inference backend while waiting
If the inference backend errors out while we wait for it to become ready
we should not wait for the timeout to expire, but rather surface the
error to the user. So change the runner.wait routine to check for the
inference backend termination and return an error in case it terminates
unexpectedly.

Signed-off-by: Piotr Stankiewicz <piotr.stankiewicz@docker.com>