Fix install race detection using more tolerant polling.
See the comments in the code for a full description, but essentially
there is a small window (which we often end up in because pull
operation deduplication synchronizes model runner installation) where a
container ID / name are registered (and we receive a "name is already in
use by container" error) before the container actually exists.

To work around this, we'll make our polling a little bit more tolerant.

This commit also replaces a string literal with a new constant.

Signed-off-by: Jacob Howard <jacob.howard@docker.com>