Log Hub response in llama-server auto update failure path
In case Hub responds with an unexpected tag (or other unexpected
content) when querying for the desired llama-server tag during the auto
update process, we return an ok error, but extra info would be nice for
debugging. So add a log entry in case the Hub response does not contain
the desired tag.

Signed-off-by: Piotr Stankiewicz <piotr.stankiewicz@docker.com>