Use more aggressive server shutdown and resequence termination
Using Shutdown with an already cancelled context will cause the method
to return almost immediately, with only idle connections being closed.
More problematically, it can't close active connections, which can
remain in flight indefinitely. At the moment, these active connections
(and their associated contexts) can cause loader.load() to block
loader.run() from exiting, especially if a backend is misbehaving, which
can cause shutdown to halt waiting on the request. Even if a backend
isn't misbehaving, an inference request can take many seconds. The
best solution would be to make loader.load() unblock if the context
passed to loader.run() is cancelled, but this is fairly complicated to
implemented. The easier solution for now is just to use a hard server
Close() to cancel inflight requests (and their contexts) and then wait
for scheduler shutdown. This is what we do in Docker Desktop.

Signed-off-by: Jacob Howard <jacob.howard@docker.com>