Add modelrunner to the video group
In order for Vulkan to work the modelrunner user needs to be a member
of the video and render groups (needed to access nodes in /dev/dri). So
add modelrunner to video when creating the user. Unfortunately render
does not have a fixed GID currently, so we'll have to deal with this
another way.

Signed-off-by: Piotr Stankiewicz <piotr.stankiewicz@docker.com>